refactor(Build): match namespace to "Contracts" folder.
