+++ src/LibraryManager.Build/Contracts/HostInteraction.cs
@@ -9,7 +9,7 @@ using System.Threading.Tasks;
 using Microsoft.Web.LibraryManager.Contracts;
 using Microsoft.Web.LibraryManager.Contracts.Configuration;
 
-namespace Microsoft.Web.LibraryManager.Build
+namespace Microsoft.Web.LibraryManager.Build.Contracts
 {
     internal class HostInteraction : IHostInteraction
     {
@@ -20,7 +20,7 @@ namespace Microsoft.Web.LibraryManager.Build
 
         public string WorkingDirectory { get; }
         public string CacheDirectory => CacheService.CacheFolder;
-        public ILogger Logger => Build.Logger.Instance;
+        public ILogger Logger => Contracts.Logger.Instance;
         public ISettings Settings => Configuration.Settings.DefaultSettings;
 
         public async Task<bool> WriteFileAsync(string path, Func<Stream> content, ILibraryInstallationState state, CancellationToken cancellationToken)
